+++ src/components/FormTask/FormTask.jsx
@@ -1,60 +1,34 @@
-import { IoClose, IoPencil, IoCheckmark } from "react-icons/io5";
-import styles from './TaskList.module.css';
+import { Field, Form, Formik } from "formik";
+import { v4 as uuidv4 } from "uuid";
+import toast, { Toaster } from 'react-hot-toast';
+import styles from './FormTask.module.css'; // Импорт стилей для FormTask
 
-const TaskList = ({
-  tasks,
-  onDelete,
-  onEdit,
-  onStatusChange,
-  editingTaskId,
-  setEditingTaskId,
-  editingText,
-  setEditingText,
-  taskStatuses,
-}) => {
+const FormTask = ({ getTask, editingTask }) => {
+  const handleSubmit = ({ task }, action) => {
+    if (!task.trim()) {
+      toast.error('Field is empty!');
+      return;
+    }
+    getTask({ task: task.trim(), id: editingTask ? editingTask.id : uuidv4() });
+    action.resetForm();
+  };
+
   return (
-    <div className={styles.taskList}>
-      {tasks?.length > 0 ? (
-        <ul className={styles.list}>
-          {tasks.map(({ task, id, status }) => (
-            <li key={id} className={styles.item}>
-              {editingTaskId === id ? (
-                <input
-                  type="text"
-                  value={editingText}
-                  onChange={(e) => setEditingText(e.target.value)}
-                  onBlur={() => onEdit(id, editingText)}
-                  onKeyPress={(e) => e.key === "Enter" && onEdit(id, editingText)}
-                  autoFocus
-                />
-              ) : (
-                <p onClick={() => { setEditingTaskId(id); setEditingText(task); }}>
-                  {task}
-                </p>
-              )}
-              <div className={styles.controls}>
-                <select
-                  value={status}
-                  onChange={(e) => onStatusChange(id, e.target.value)}
-                >
-                  {taskStatuses.map((status) => (
-                    <option key={status} value={status}>
-                      {status}
-                    </option>
-                  ))}
-                </select>
-                <button onClick={() => onDelete(id)} className={styles.deleteButton}>
-                  <IoClose size={20} />
-                </button>
-              </div>
-            </li>
-          ))}
-        </ul>
-      ) : (
-        <p className={styles.emptyMessage}>Список задач пуст! 😥</p>
-      )}
-    </div>
+    <>
+      <Formik
+        initialValues={{ task: editingTask ? editingTask.task : "" }}
+        onSubmit={handleSubmit}
+      >
+        <Form className={styles.form}>
+          <Field name="task" placeholder="Your task ..." className={styles.input} />
+          <button type="submit" className={styles.button}>
+            {editingTask ? "Update" : "Add"}
+          </button>
+        </Form>
+      </Formik>
+      <Toaster />
+    </>
   );
 };
 
-export default TaskList;
+export default FormTask;
