+++ src/components/FormTask/FormTask.module.css
@@ -23,16 +23,4 @@
   .button:hover {
     background-color: #0056b3;
   }
-
-  .editButton {
-    background: none;
-    border: none;
-    cursor: pointer;
-    color: #007bff;
-    margin-right: 10px;
-  }
-  
-  .editButton:hover {
-    color: #0056b3;
-  }
