+++ src/components/TaskList/TaskList.jsx
@@ -1,18 +1,48 @@
 import { IoClose, IoPencil } from "react-icons/io5";
 import styles from './TaskList.module.css';
 
-const TaskList = ({ tasks, onDelete, onEdit }) => {
+const TaskList = ({
+  tasks,
+  onDelete,
+  onEdit,
+  onStatusChange,
+  editingTaskId,
+  setEditingTaskId,
+  editingText,
+  setEditingText,
+  taskStatuses,
+}) => {
   return (
-    <>
+    <div className={styles.taskList}>
       {tasks?.length > 0 ? (
         <ul className={styles.list}>
-          {tasks.map(({ task, id }) => (
+          {tasks.map(({ task, id, status }) => (
             <li key={id} className={styles.item}>
-              <p>{task}</p>
-              <div>
-                <button onClick={() => onEdit(id)} className={styles.editButton}>
-                  <IoPencil size={16} />
-                </button>
+              {editingTaskId === id ? (
+                <input
+                  type="text"
+                  value={editingText}
+                  onChange={(e) => setEditingText(e.target.value)}
+                  onBlur={() => onEdit(id, editingText)}
+                  onKeyPress={(e) => e.key === "Enter" && onEdit(id, editingText)}
+                  autoFocus
+                />
+              ) : (
+                <p onClick={() => { setEditingTaskId(id); setEditingText(task); }}>
+                  {task}
+                </p>
+              )}
+              <div className={styles.controls}>
+                <select
+                  value={status}
+                  onChange={(e) => onStatusChange(id, e.target.value)}
+                >
+                  {taskStatuses.map((status) => (
+                    <option key={status} value={status}>
+                      {status}
+                    </option>
+                  ))}
+                </select>
                 <button onClick={() => onDelete(id)} className={styles.deleteButton}>
                   <IoClose size={20} />
                 </button>
@@ -21,9 +51,9 @@ const TaskList = ({ tasks, onDelete, onEdit }) => {
           ))}
         </ul>
       ) : (
-        <p className={styles.emptyMessage}>List is empty! 😥</p>
+        <p className={styles.emptyMessage}>Список задач пуст! 😥</p>
       )}
-    </>
+    </div>
   );
 };
