+++ src/components/TaskList/TaskList.module.css
@@ -1,4 +1,8 @@
-.list {
+.taskList {
+    margin-top: 20px;
+  }
+  
+  .list {
     list-style: none;
     padding: 0;
   }
@@ -13,6 +17,12 @@
     margin-bottom: 10px;
   }
   
+  .controls {
+    display: flex;
+    gap: 10px;
+    align-items: center;
+  }
+  
   .deleteButton {
     background: none;
     border: none;
@@ -20,7 +30,15 @@
     color: #ff4d4d;
   }
   
+  .editButton {
+    background: none;
+    border: none;
+    cursor: pointer;
+    color: #007bff;
+  }
+  
   .emptyMessage {
     text-align: center;
     color: #666;
   }
+
